Reinstate `auth` endpoints
