+++ src/models/Admin.js
@@ -5,17 +5,20 @@ const AdminSchema = new mongoose.Schema({
   name: { type: String, required: true },
   email: { type: String, required: true, unique: true },
   password: { type: String, required: true },
-  role: { type: String, default: 'admin', immutable: true },  // Make role immutable means it can not be changes
+  role: { type: String, default: 'admin', immutable: true },  // Make role immutable means it can not be changed
   lastPasswordChange: { type: Date, default: Date.now },
+  failedLoginAttempts: { type: Number, default: 0 },
   created_at: { type: Date, default: Date.now },
   updated_at: { type: Date, default: Date.now }
 });
 
 // Hash the password before saving
 AdminSchema.pre('save', async function (next) {
+  this.updated_at = Date.now();
   if (!this.isModified('password')) return next();
   const salt = await bcrypt.genSalt(10);
   this.password = await bcrypt.hash(this.password, salt);
+  this.lastPasswordChange = Date.now();
   next();
 });
