+++ src/models/User.js
@@ -1,17 +1,23 @@
 const mongoose = require('mongoose');
+const bcrypt = require('bcryptjs');
 
 const UserSchema = new mongoose.Schema({
-  username: { type: String, required: true, unique: true },
+  fullname: { type: String, required: true },
   email: { type: String, required: true, unique: true },
   password_hash: { type: String, required: true },
+  role_id: { type: mongoose.Schema.Types.ObjectId, ref: 'UserRole' },
   lastPasswordChange: { type: Date, default: Date.now },
   failedLoginAttempts: { type: Number, default: 0 },
   created_at: { type: Date, default: Date.now },
   updated_at: { type: Date, default: Date.now }
 });
 
-UserSchema.pre('save', function (next) {
+UserSchema.pre('save', async function (next) {
   this.updated_at = Date.now();
+  if (!this.isModified('password_hash')) return next();
+  this.lastPasswordChange = Date.now();
+  const salt = await bcrypt.genSalt(10);
+  this.password_hash = await bcrypt.hash(this.password_hash, salt);
   next();
 });
