+++ src/server.js
@@ -46,12 +46,14 @@ const wifiCSIRoutes = require('./routes/wifiCSI');
 const activityRecognitionRoutes = require('./routes/activityRecognition');
 const alertsRoutes = require('./routes/alerts');
 const nurseRoutes = require('./routes/nurseRoutes'); // Import nurse routes
+const userRoutes = require('./routes/user');
 
+app.use('/swaggerDocs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
 app.use('/api/v1/wifi-csi', wifiCSIRoutes);
 app.use('/api/v1/activity-recognition', activityRecognitionRoutes);
-app.use('/swaggerDocs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
 app.use('/api/v1/alerts', alertsRoutes);
-app.use("/api/v2/nurse", nurseRoutes);
+app.use("/api/v1/nurse", nurseRoutes);
+app.use("/api/v1/auth", userRoutes);
 
 app.get('/redoc', (req, res) => {
   res.send(`
